Issue #689 - Remove layer of SettingsChannel, use Adapter directly
